refactor(transport): move peer capability learning to discovery-tags module
Move the PeerCapabilities interface and learnPeerCapabilities function from the
separate capability-learner.ts file into discovery-tags.ts. This consolidates
all discovery tag parsing logic in one place and simplifies the capability
detection by handling the oversizedEnabled check in the caller. The
capability-learner.ts file is now redundant and has been deleted.
